implement terminal window size change notification
descriptor.h, interface.c:
change all window size and coordinate variables to ushort since that's what struct winsize uses

interface.c:change the stored window size and send SIGWINCH to the foreground process group in ltm_set_window_dimensions
init.c:change the stored window size in ltm_term_init since it can't happen before that (since there's no pty)

interface.c:in ltm_set_window_dimensions, first check whether the new width and height are the same as the old ones and exit immediately if they are

interface.c:use a config.h macro to determine whether <sys/ioctl.h> need be included

+++ src/descriptor.h
@@ -14,8 +14,8 @@
 #define ALTSCREEN 1
 
 struct point {
-	uint x;
-	uint y;
+	ushort x;
+	ushort y;
 };
 
 struct area {
@@ -40,8 +40,8 @@ struct ltm_term_desc {
 	char * shell;
 	pid_t pid;
 
-	uint width;
-	uint height;
+	ushort width;
+	ushort height;
 
 	struct point cursor;
